requirements.txt
streamlit>=1.18.0
tensorflow>=2.9.0
numpy>=1.21.0
Pillow>=9.0.0
